requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.9
django-braces==1.15.0
django-model-utils==4.2.0
djangorestframework==3.12.4
idna==3.3
psycopg2-binary==2.9.2
python-decouple==3.5
pytz==2021.3
requests==2.26.0
sqlparse==0.4.2
swapper==1.2.0
urllib3==1.26.7
